// packages/simmerbase/src/schemas/trap_types.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const TrapTypesBaseRowSchema = z.object({
	created_at: z.coerce.date<Date>(),
	created_by: z.uuid().nullable(),
	group_id: z.uuid().nullable(),
	id: z.uuid(),
	shorthand: z.string().nullable(),
	trap_type_name: z.string(),
	updated_at: z.coerce.date<Date>(),
	updated_by: z.uuid().nullable(),
});

export type TrapTypesBaseRow = z.infer<typeof TrapTypesBaseRowSchema>;

export const TrapTypesInsertSchema = z.object({
	group_id: z.uuid().nullable(),
	id: z.uuid(),
	shorthand: z.string().nullable(),
	trap_type_name: z.string(),
});

export type TrapTypesInsert = z.infer<typeof TrapTypesInsertSchema>;

export const TrapTypesUpdateSchema = TrapTypesInsertSchema.omit({ id: true }).partial();

export type TrapTypesUpdate = z.infer<typeof TrapTypesUpdateSchema>;